[Chore] Move bucketlist service.
