+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 
 import { AppComponent }  from './app.component';
-import  { BucketlistListComponent } from './bucketlist/bucketlist-list.component';
+import  { BucketlistListComponent } from './bucketlist/buckets/bucketlist-list.component';
 import { BucketlistService } from './bucketlist/all-bucketlists.service';
 
 // services
@@ -13,7 +13,7 @@ import { AuthenticationService } from './user/authentication.service';
 
 // components
 import { LoginComponent } from './user/login/login.component';
-import { ItemComponent } from './bucketlist/item.component';
+import { ItemComponent } from './bucketlist/buckets/item.component';
 
 const appRoutes: Routes = [
   {
